PySide6 -> PyQt6, fix window switching problem
PySide6 and PyQt6 are pretty much the same.

Switching to PySide6 broke window switching so I switched back to PyQt6.

The issue with PyQt6 was that QImage.Format_Grayscale16 wasn't working (but does work in PySide6).

In PyQt6, it becomes QImage.Format.Format_Grayscale16. Minor difference but was a headache to figure out.