Added code for debugging, see description
Every time a re-render of a slice occurs, the `sitk.Image` and `np.array` data types (usually `int16` except the top file, which is `float32`) will be printed to the terminal.

The min and max values of the `np` array are also printed.

Lastly, all values in the `np` array are written to `src/GUI/np_array_slice.txt` (gitignored). Note this does slow down performance. Without that writing, the rendering is really fast.