Circumference window displays contour. Menu bar bugged.
Found the bug causing File > Export > CSV to not be enabled when switching to circumference window.

When switching to circumference window, the menu bar doesn't actually change to the circumference window menu bar.

I added a GUI element to main.ui but not circumference.ui, yet that GUI element showed up on circumference window.

Also, I am 100% certain that exporting as PNG in circumference window should export the slice with a contour. However, it exports the slice without a contour, meaning the menu bar item must be that of the main window, not the circumference window.

This might be caused by how I set up the stacked widget in main, not in QtDesigner.