Basic, untested arc length with pixel spacing.
imgproc.curr_arc_length is from a previous commit in which the function was unit tested and exactly matched the result of cv2.arcLength given identical inputs.

However, it now accounts for pixel spacing so won't be exactly the same.

Also, not 100% sure x and y are in the right order because of transposing issues.