Images oriented for Z view on load, refactoring
Orienting for Z view on load fixes a bug with the misaligned image.

The misaligned image's slice slider had the wrong maximum value since it was GetSize()[2], but this wasn't the inferior-superior axis. This is solved by orienting it to the Z view, which GetSize()[2] to be the inferior-superior axis.

Also other minor refactoring stuff.